+++ src/packet/error.rs
@@ -19,6 +19,7 @@ pub enum PacketError {
     ErrVP8CorruptedPacket,
     ErrVP9CorruptedPacket,
     ErrAv1CorruptedPacket,
+    ErrRedCorruptedPacket,
 }
 
 impl fmt::Display for PacketError {
@@ -45,6 +46,7 @@ impl fmt::Display for PacketError {
             PacketError::ErrVP8CorruptedPacket => write!(f, "VP8 corrupted packet"),
             PacketError::ErrVP9CorruptedPacket => write!(f, "VP9 corrupted packet"),
             PacketError::ErrAv1CorruptedPacket => write!(f, "AV1 corrupted packet"),
+            PacketError::ErrRedCorruptedPacket => write!(f, "RED corrupted packet"),
         }
     }
 }